+++ enablebanking_cli/commands/app.py
@@ -32,20 +32,109 @@ class AppCommand(BaseCommand):
         self.subparsers = self.parser.add_subparsers(
             title="Application Commands",
             dest="app_command")
-        register_parser = self.subparsers.add_parser("register", help="Register an application")
-        register_parser.add_argument(
-            "-n",
-            "--name",
-            type=str,
-            help="Name of the application being registered",
-            required=True,
+        default_parser = self.subparsers.add_parser(
+            "default",
+            help="Set an application to be used by default",
         )
+        default_parser.add_argument("app", type=str, help="Application ID")
+        list_parser = self.subparsers.add_parser("list", help="List locally available applications")
+        requests_parser = self.subparsers.add_parser(
+            "requests",
+            help="Fetch logs of requests made by an application",
+        )
+        requests_parser.add_argument(
+            "-a",
+            "--app",
+            type=str,
+            help="Application ID (using default if not provided)",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--account-id",
+            type=str,
+            help="Filtering requests associated with an account by its ID",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--aspsp-country",
+            type=str,
+            help="Filtering requests by a country (ISO 3166 code)",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--aspsp-name",
+            type=str,
+            help="Filtering requests by ASPSP name",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--auth-approach",
+            type=str,
+            help="Filtering requests by an authorization approach",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--auth-method",
+            type=str,
+            help="Filtering requests by an authorization method",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--authorization-id",
+            type=str,
+            help="Filtering requests associated with an authorization by its ID",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--endpoint-name",
+            type=str,
+            help="Filtering requests by an endpoint name",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--payment-id",
+            type=str,
+            help="Filtering requests associated with a payment by its ID",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--psu-type",
+            type=str,
+            help="Filtering requests by PSU type",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--response-code",
+            type=str,
+            help="Filtering requests by a response code",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--session-id",
+            type=str,
+            help="Filtering requests associated with a session by its ID",
+            required=False,
+        )
+        requests_parser.add_argument(
+            "--session-status",
+            type=str,
+            help="Filtering requests by a session status",
+            required=False,
+        )
+        register_parser = self.subparsers.add_parser("register", help="Register an application")
         register_parser.add_argument(
             "-e",
             "--environment",
             type=Environment,
             choices=[*Environment],
             help=f"Environment, which the application will use",
+            required=True,
+        )
+        register_parser.add_argument(
+            "-n",
+            "--name",
+            type=str,
+            help="Name of the application being registered",
             required=True,
         )
         register_parser.add_argument(
@@ -57,10 +146,43 @@ class AppCommand(BaseCommand):
             required=True,
         )
         register_parser.add_argument(
+            "-d",
+            "--description",
+            type=str,
+            help="Description of the application being registered",
+            required=False,
+        )
+        register_parser.add_argument(
+            "--gdpr-email",
+            type=str,
+            help=f"Email of data protection matters",
+            required=False,
+        )
+        register_parser.add_argument(
+            "--privacy-url",
+            type=str,
+            help=f"URL of the application's privacy policy",
+            required=False,
+        )
+        register_parser.add_argument(
+            "--terms-url",
+            type=str,
+            help=f"URL of the application's terms of service",
+            required=False,
+        )
+        register_parser.add_argument(
+            "-c",
             "--cert-path",
             type=str,
-            help=f"Redirect URL(s) allowed for the application",
+            help=f"Path to a certificate or a public key of the application",
             required=True,
+        )
+        register_parser.add_argument(
+            "-k",
+            "--key-path",
+            type=str,
+            help=f"Path to a private key of the application (used to generate or verify public key)",
+            required=False,
         )
         self.subparsers.add_parser("share", help="Share an application")
 
@@ -75,9 +197,77 @@ class AppCommand(BaseCommand):
             "certificate": cert_content,
             "environment": args.environment,
             "redirect_urls": args.redirect_urls,
+            "description": args.description,
+            "gdpr_email": args.gdpr_email,
+            "privacy_url": args.privacy_url,
+            "terms_url": args.terms_url,
         })
         if response.status != 200:
             print(f"{response.status} response from the applications API: {response.read().decode()}")
             return 1
         response_data = json.loads(response.read().decode())
         print(f"The application is registered under ID {response_data['app_id']}")
+        cp_store = CpStore(args.root_path)
+        os.makedirs(cp_store.cp_apps_path, exist_ok=True)
+        app_filename = cp_store.get_app_filename(response_data["app_id"])
+        with open(os.path.join(cp_store.cp_apps_path, app_filename), "w") as f:
+            f.write(json.dumps({
+                "kid": response_data["app_id"],
+                "name": args.name,
+                "description": args.description,
+                "certificate": cert_content,
+                "key_path": args.key_path,
+                "environment": args.environment,
+                "redirect_urls": args.redirect_urls,
+            }, indent=4))
+        cp_store.set_default_app_filename(app_filename)
+        print("Done!")
+
+    def list(self, args):
+        cp_store = CpStore(args.root_path)
+        apps_data = cp_store.load_app_files()
+        for app_data in apps_data:
+            print("*" if app_data["_default"] else " ", app_data["kid"], app_data["name"])
+
+    def default(self, args):
+        cp_store = CpStore(args.root_path)
+        apps_data = cp_store.load_app_files()
+        is_app_found = False
+        for app_data in apps_data:
+            if app_data["kid"] == args.app:
+                is_app_found = True
+                break
+        if not is_app_found:
+            print(f"Application with ID '{args.app}' is not available")
+            return 1
+        cp_store.set_default_app_filename(cp_store.get_app_filename(args.app))
+        print(f"Default application switched to {args.app} ({app_data['name']})")
+
+    def requests(self, args):
+        cp_store = CpStore(args.root_path)
+        cp_client = CpClient(args.cp_domain, cp_store.get_user_path(args.user))
+        app_id = args.app if args.app is not None else cp_store.load_app_file()["kid"]
+        print("Fetching requests...")
+        response = cp_client.fetch_requests(
+            app_id,
+            account_id=args.account_id,
+            aspsp_country=args.aspsp_country,
+            aspsp_name=args.aspsp_name,
+            auth_approach=args.auth_approach,
+            auth_method=args.auth_method,
+            authorization_id=args.authorization_id,
+            endpoint_name=args.endpoint_name,
+            payment_id=args.payment_id,
+            psu_type=args.psu_type,
+            response_code=args.response_code,
+            session_id=args.session_id,
+            session_status=args.session_status,
+        )
+        if response.status != 200:
+            print(f"{response.status} response from the requests API: {response.read().decode()}")
+            return 1
+        response_data = json.loads(response.read().decode())
+        requests = response_data["requests"]
+        print(f"Fetched {len(requests)} requests" + (":" if len(requests) else "."))
+        print(json.dumps(requests, indent=4))
+        print("Done!")