+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+- Added the `app requests` command for fetching the log of requests made by an application.
+- Made the `app register` command to store application properties locally, so that they are easy to
+  use with other commands.
+- Added optional parameters for the `app register` command: `-d/--description`, `--gdpr-email`,
+  `--privacy-url` and `--terms-url`.
+- Small fixes and improvements.
+
 ## 0.1.1 - 2024-01-10
 
 Refreshing token on the response code 401 instead of 403 as the control panel API has changed